Refactored replay.js to facilitate easier testing/readability
