+++ requestIdBuilder.js
@@ -1,0 +1,1 @@
+export default request => `${request.method} ${request.url}`;
