+++ requestRepeatMapBuilder.js
@@ -1,0 +1,20 @@
+import buildRequestId from './requestIdBuilder';
+
+export default (requests) => {
+  const repeatMap = [];
+
+  requests.forEach(({ request }) => {
+    const requestId = buildRequestId(request);
+
+    if (requestId in repeatMap) {
+      repeatMap[requestId].repeated += 1;
+    } else {
+      repeatMap[requestId] = {
+        repeated: 1,
+        invocations: 0,
+      };
+    }
+  });
+
+  return repeatMap;
+};
